Neighbs layer toggle: move to /Explore/Neighborhood/Something intro
