+++ src/components/explore/index.ts
@@ -6,6 +6,7 @@ export * from './Explore'
 export * from './ExploreLanding'
 export * from './LangCardsList'
 export * from './LangProfileDescrip'
+export * from './LayerToggle'
 export * from './MidLevelExplore'
 export * from './NeighborhoodsInstance'
 export * from './NeighborhoodsLanding'
